+++ VMagicMirror/Assets/Baku/VMagicMirror/Scripts/AvatarControl/FaceContol/EyeBoneAngleSetter.cs
@@ -1,5 +1,6 @@
 using Baku.VMagicMirror.IK;
 using UnityEngine;
+using VRM;
 using Zenject;
 
 namespace Baku.VMagicMirror
@@ -13,6 +14,11 @@ namespace Baku.VMagicMirror
     /// </remarks>
     public class EyeBoneAngleSetter : MonoBehaviour
     {
+        private static readonly BlendShapeKey LookLeftKey = BlendShapeKey.CreateFromPreset(BlendShapePreset.LookLeft);
+        private static readonly BlendShapeKey LookRightKey = BlendShapeKey.CreateFromPreset(BlendShapePreset.LookRight);
+        private static readonly BlendShapeKey LookUpKey = BlendShapeKey.CreateFromPreset(BlendShapePreset.LookUp);
+        private static readonly BlendShapeKey LookDownKey = BlendShapeKey.CreateFromPreset(BlendShapePreset.LookDown);
+        
         //NOTE: 実験した範囲ではこのリミットを超えることは滅多にない
         private const float RateMagnitudeLimit = 1.2f;
         
@@ -29,17 +35,22 @@ namespace Baku.VMagicMirror
         [SerializeField] private EyeDownMotionController eyeDownMotionController;
         [SerializeField] private EyeJitter eyeJitter;
         [SerializeField] private ExternalTrackerEyeJitter externalTrackerEyeJitter;
+        [SerializeField] private BehaviorBasedAutoBlinkAdjust blinkAdjust;
         
         private NonImageBasedMotion _nonImageBasedMotion;
-        private EyeBoneAngleMapApplier _angleMapApplier;
+        private EyeBoneAngleMapApplier _boneApplier;
+        private EyeBlendShapeMapApplier _blendShapeApplier;
         private EyeLookAt _eyeLookAt;
+        //NOTE: BlendShapeResultSetterの更に後処理として呼び出す(ホントは前処理で値を入れたいが、Execution Order的に難しい)
+        private VRMBlendShapeProxy _blendShape = null;
 
         private Transform _leftEye;
         private Transform _rightEye;
         private bool _hasModel;
-        private bool _hasLeftEye;
-        private bool _hasRightEye;
+        private bool _hasLeftEyeBone;
+        private bool _hasRightEyeBone;
 
+        private bool _moveEyesDuringFaceClipApplied = false;
         private float _motionScale = 1f;
         private float _motionScaleWithMap = 1f;
         private bool _useAvatarEyeCurveMap = true;
@@ -56,11 +67,17 @@ namespace Baku.VMagicMirror
             IKTargetTransforms ikTargets,　NonImageBasedMotion nonImageBasedMotion)
         {
             _nonImageBasedMotion = nonImageBasedMotion;
-            _angleMapApplier = new EyeBoneAngleMapApplier(vrmLoadable);
+            _boneApplier = new EyeBoneAngleMapApplier(vrmLoadable);
+            _blendShapeApplier = new EyeBlendShapeMapApplier(vrmLoadable);
             _eyeLookAt = new EyeLookAt(vrmLoadable, ikTargets.LookAt);
 
             vrmLoadable.VrmLoaded += OnVrmLoaded;
             vrmLoadable.VrmDisposing += OnVrmDisposed;
+            
+            receiver.AssignCommandHandler(
+                VmmCommands.EnableEyeMotionDuringClipApplied, 
+                command => _moveEyesDuringFaceClipApplied = command.ToBoolean()
+                );
             
             receiver.AssignCommandHandler(
                 VmmCommands.SetEyeBoneRotationScale,
@@ -83,19 +100,20 @@ namespace Baku.VMagicMirror
         {
             _leftEye = info.animator.GetBoneTransform(HumanBodyBones.LeftEye);
             _rightEye = info.animator.GetBoneTransform(HumanBodyBones.RightEye);
+            _blendShape = info.blendShape;
 
-            //NOTE: せっかく整頓しているので、片目だけボーンがあるモデルでちゃんと動くことを検討する
-            _hasLeftEye = _leftEye != null;
-            _hasRightEye = _rightEye != null;
+            _hasLeftEyeBone = _leftEye != null;
+            _hasRightEyeBone = _rightEye != null;
             _hasModel = true;
         }
 
         private void OnVrmDisposed()
         {
             _hasModel = false;
-            _hasLeftEye = false;
-            _hasRightEye = false;
+            _hasLeftEyeBone = false;
+            _hasRightEyeBone = false;
 
+            _blendShape = null;
             _leftEye = null;
             _rightEye = null;
         }
@@ -114,7 +132,8 @@ namespace Baku.VMagicMirror
         
         private void LateUpdate()
         {
-            if (!_hasModel || (!_hasLeftEye && !_hasRightEye))
+            if (!_hasModel || 
+                (_boneApplier.HasLookAtBoneApplier && !_hasLeftEyeBone && !_hasRightEyeBone))
             {
                 return;
             }
@@ -137,6 +156,12 @@ namespace Baku.VMagicMirror
             leftRate = Vector2.ClampMagnitude(leftRate, RateMagnitudeLimit);
             rightRate = Vector2.ClampMagnitude(rightRate, RateMagnitudeLimit);
 
+            var meanRateWithoutJitter = 0.5f * (
+                (leftRate + rightRate) - 
+                (eyeJitter.LeftEyeRotationRate + eyeJitter.RightEyeRotationRate)
+            );
+            blinkAdjust.SetEyeMoveRate(meanRateWithoutJitter);
+
             // 符号に注意、Unityの普通の座標系ではピッチは下が正
             var leftYaw = leftRate.x * HorizontalRateToAngle;
             var leftPitch = -leftRate.y * VerticalRateToAngle;
@@ -150,42 +175,60 @@ namespace Baku.VMagicMirror
             leftPitch += _eyeLookAt.Pitch;
             rightPitch += _eyeLookAt.Pitch;
 
-            if (ReserveReset)
+            if (ReserveReset && !_moveEyesDuringFaceClipApplied)
             {
                 //NOTE: 0でもmap処理が入った結果、非ゼロの角度が入る可能性があるのでreturnはしない
                 leftPitch = 0f;
                 leftYaw = 0f;
                 rightPitch = 0f;
                 rightYaw = 0f;
-                ReserveReset = false;
             }
             else
             {
                 var motionScale = _useAvatarEyeCurveMap
                     ? _motionScaleWithMap
                     : _motionScale * factorWhenMapDisable;
-                var weightFactor = motionScale * ReserveWeight;
+                var weightFactor = motionScale * (_moveEyesDuringFaceClipApplied ? 1 : ReserveWeight);
                 leftPitch = ScaleAndClampAngle(leftPitch, weightFactor);
                 leftYaw = ScaleAndClampAngle(leftYaw, weightFactor);
                 rightPitch = ScaleAndClampAngle(rightPitch, weightFactor);
                 rightYaw = ScaleAndClampAngle(rightYaw, weightFactor);
             }
+            ReserveReset = false;
             ReserveWeight = 1f;
-            
-            if (_useAvatarEyeCurveMap)
-            {
-                (leftYaw, leftPitch) = _angleMapApplier.GetLeftMappedValues(leftYaw, leftPitch);
-                (rightYaw, rightPitch) = _angleMapApplier.GetRightMappedValues(rightYaw, rightPitch);
-            }
 
-            if (_hasLeftEye)
+            if (_boneApplier.HasLookAtBoneApplier)
             {
-                _leftEye.localRotation = Quaternion.Euler(leftPitch, leftYaw, 0f);
-            }
+                if (_useAvatarEyeCurveMap)
+                {
+                    (leftYaw, leftPitch) = _boneApplier.GetLeftMappedValues(leftYaw, leftPitch);
+                    (rightYaw, rightPitch) = _boneApplier.GetRightMappedValues(rightYaw, rightPitch);
+                }
 
-            if (_hasRightEye)
+                if (_hasLeftEyeBone)
+                {
+                    _leftEye.localRotation = Quaternion.Euler(leftPitch, leftYaw, 0f);
+                }
+
+                if (_hasRightEyeBone)
+                {
+                    _rightEye.localRotation = Quaternion.Euler(rightPitch, rightYaw, 0f);
+                }                
+            }
+            else
             {
-                _rightEye.localRotation = Quaternion.Euler(rightPitch, rightYaw, 0f);
+                //elseの場合はBlendShapeの処理に落ち、このケースは強制的にアバターの目カーブを参照する
+                //(参照しないと基準も何もないので)
+                var result = _blendShapeApplier.GetMappedValues(
+                    0.5f * (leftYaw + rightYaw),
+                    0.5f * (leftPitch + rightPitch)
+                );
+                
+                _blendShape.AccumulateValue(LookLeftKey, result.Left);
+                _blendShape.AccumulateValue(LookRightKey, result.Right);
+                _blendShape.AccumulateValue(LookUpKey, result.Up);
+                _blendShape.AccumulateValue(LookDownKey, result.Down);
+                _blendShape.Apply();
             }
         }