+++ WPF/VMagicMirrorConfig/Model/Entity/MotionSetting.cs
@@ -63,6 +63,7 @@
         public bool UseLookAtPointMousePointer { get; set; } = true;
         public bool UseLookAtPointMainCamera { get; set; } = false;
 
+        public bool MoveEyesDuringFaceClipApplied { get; set; } = false;
         public bool UseAvatarEyeBoneMap { get; set; } = true;
         public int EyeBoneRotationScale { get; set; } = 100;
         public int EyeBoneRotationScaleWithMap { get; set; } = 100;
@@ -77,6 +78,8 @@
 
         //NOTE: dB単位なので0がデフォルト。対数ベースのほうがレンジ取りやすい
         public int MicrophoneSensitivity { get; set; } = 0;
+
+        public bool AdjustLipSyncByVolume { get; set; } = true;
 
         #endregion
 
@@ -131,6 +134,7 @@
             EnableLipSync = true;
             LipSyncMicrophoneDeviceName = "";
             MicrophoneSensitivity = 0;
+            AdjustLipSyncByVolume = true;
         }
 
         public void ResetFaceEyeSetting()
@@ -140,6 +144,7 @@
             UseLookAtPointMousePointer = true;
             UseLookAtPointMainCamera = false;
 
+            MoveEyesDuringFaceClipApplied = false;
             UseAvatarEyeBoneMap = true;
             EyeBoneRotationScale = 100;
             EyeBoneRotationScaleWithMap = 100;